Refactor: Enhance tenant-aware handling across various repositories and use cases; update method signatures to include TenantId for improved data management

+++ custom-domain/source/uim/platform/custom_domain/domain/entities/domain_dashboard.d
@@ -32,7 +32,7 @@ struct CertificateExpirationWarning {
     string certificateId;
     string domainName;
     long expiresAt;
-    int daysRemaining;
+    long daysRemaining;
     ExpirationSeverity severity;
 
     Json toJson() const {
